+++ src/install.js
@@ -1,0 +1,4 @@
+import { activatePlugin } from 'kea'
+import thunkPlugin from './index'
+
+activatePlugin(thunkPlugin)